photo_manager_cli/requirements-dev.txt
black
mypy
isort
pylint
pytest
coverage
vcrpy
